feat : refactoring (#10)
