+++ src/main/java/deu/cse/spring_webmail/model/SHAPasswordAlgorithm.java
@@ -2,10 +2,8 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package deu.cse.spring_webmail.unittest;
+package deu.cse.spring_webmail.model;
 
-import deu.cse.spring_webmail.model.SHAPasswordAlgorithm;
-import deu.cse.spring_webmail.model.UserService;
 import jakarta.mail.MessagingException;
 import jakarta.mail.internet.MimeUtility;
 import java.io.ByteArrayOutputStream;
@@ -14,56 +12,28 @@ import java.io.OutputStream;
 import java.io.UnsupportedEncodingException;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 import lombok.extern.slf4j.Slf4j;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 
 /**
  *
  * @author 정기석
  */
 @Slf4j
-@SpringBootTest
-public class SHAAlgorithmTest {
-       
-    @Autowired SHAPasswordAlgorithm SHAPasswordAlgorithm;
+public class SHAPasswordAlgorithm {
     
-    @Test
-    void sha(){
-        String a = SHAPasswordAlgorithm.createPwdHash();
-        log.info("a = {}",a);
-    }
-    
-    
-    @Test
-    void 이메일자르기테스트(){
-        String email = "abce@naver.com";
-        String result = email.substring(0,email.lastIndexOf("@"));
-        log.info("result = {}",result);
-    }
-    
-    
-    @Test
-    void shaTest(){
+    public String createPwdHash(){
+        String result = "";
         try {
-            //given
-            String password = "admin";
-            String algorithem = "SHA";
-            //when
-            String result = digestString(password, algorithem);
-             //then
-            log.info("result = {}",result);
+            result = digestString("1234","SHA");
+            return result;
         } catch (NoSuchAlgorithmException ex) {
             log.error("exception = {}",ex);
-                
         } catch (UnsupportedEncodingException ex) {
-            log.error("exception = {}",ex);
+            log.error("exception = {}",ex);         
         }
-       
+        return result;
     }
+    
     private String digestString(String pass, String algorithm ) throws NoSuchAlgorithmException, UnsupportedEncodingException  {
 
         MessageDigest md;
@@ -71,6 +41,7 @@ public class SHAAlgorithmTest {
 
         try {
             md = MessageDigest.getInstance(algorithm);
+            log.info("md = {}",md.toString());
             byte[] digest = md.digest(pass.getBytes("iso-8859-1"));
             bos = new ByteArrayOutputStream();
             OutputStream encodedStream = MimeUtility.encode(bos, "base64");
